requirements.txt
yt-dlp
requests
python-dotenv
loguru